+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## NEXT
 
+  * Add `--no-dev` and `--optimize-autoloader` to default composer install command.
   * Supports new ey.yml options: `bundler`, `composer`, and `npm` which can be set to true or false to enable or disable each dependency manager.
 
 ## v2.3.0 (2013-08-07)