+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## NEXT
 
+  * Fix find maxdepth problem that causes shared/config files to by symlinked incorrectly.
   * Add `--no-dev` and `--optimize-autoloader` to default composer install command.
   * Supports new ey.yml options: `bundler`, `composer`, and `npm` which can be set to true or false to enable or disable each dependency manager.
   * Supports experimental ey.yml option: `restart_groups: n`. When set to a number `n` greater than 1, app servers are divided into `n` groups and each group is restarted serially.